factor out info modal component and move stuff into components folder
