+++ src/App.js
@@ -1,15 +1,15 @@
 import { useEffect, useState, useRef } from 'react'
 import { letters, status } from './constants'
-import { Keyboard } from './Keyboard'
+import { Keyboard } from './components/Keyboard'
 import answers from './data/answers'
 import words from './data/words'
 import Modal from 'react-modal'
 import Success from './data/Success.png'
 import Fail from './data/Cross.png'
 import { useLocalStorage } from './hooks/useLocalStorage'
-import { ReactComponent as Github } from './data/Github.svg'
-import { ReactComponent as Close } from './data/Close.svg'
+
 import { ReactComponent as Info } from './data/Info.svg'
+import { InfoModal } from './components/InfoModal'
 
 Modal.setAppElement('#root')
 
@@ -287,7 +287,7 @@ function App() {
     )
   }
 
-  const customStyles = {
+  const modalStyles = {
     overlay: {
       position: 'fixed',
       top: 0,
@@ -387,10 +387,17 @@ function App() {
           )}
           <PlayAgainButton />
         </div>
+        <InfoModal
+          isOpen={infoModalIsOpen}
+          handleClose={handleInfoClose}
+          toggleDarkMode={toggleDarkMode}
+          darkMode={darkMode}
+          styles={modalStyles}
+        />
         <Modal
           isOpen={modalIsOpen}
           onRequestClose={closeModal}
-          style={customStyles}
+          style={modalStyles}
           contentLabel="Game End Modal"
         >
           <div className={darkMode ? 'dark' : ''}>
@@ -435,72 +442,6 @@ function App() {
           onDeletePress={onDeletePress}
           gameDisabled={gameState !== state.playing}
         />
-        <Modal
-          isOpen={infoModalIsOpen}
-          onRequestClose={handleInfoClose}
-          style={customStyles}
-          contentLabel="Game Info Modal"
-        >
-          <div className={`h-full ${darkMode ? 'dark' : ''}`}>
-            <button
-              className="absolute top-4 right-4 rounded-full nm-flat-background dark:nm-flat-background-dark text-primary dark:text-primary-dark p-1 w-6 h-6 sm:p-2 sm:h-8 sm:w-8"
-              onClick={handleInfoClose}
-            >
-              <Close />
-            </button>
-            <div className="h-full flex flex-col items-center justify-center max-w-[390px] mx-auto pt-9 text-primary dark:text-primary-dark">
-              <div className="flex-1 w-full sm:text-base text-sm">
-                <h1 className="text-center sm:text-3xl text-2xl">How to play</h1>
-                <ul className="list-disc pl-5 block sm:text-base text-sm">
-                  <li className="mt-6 mb-2">You have 6 guesses to guess the correct word.</li>
-                  <li className="mb-2">You can guess any valid word.</li>
-                  <li className="mb-2">
-                    After each guess, each letter will turn green, yellow, or gray.
-                  </li>
-                </ul>
-                <div className="mb-3 mt-8 flex items-center">
-                  <span className="nm-inset-n-green text-gray-50 inline-flex items-center justify-center text-3x w-10 h-10 rounded-full">
-                    W
-                  </span>
-                  <span className="mx-2">=</span>
-                  <span>Correct letter, correct spot</span>
-                </div>
-                <div className="mb-3">
-                  <span className="nm-inset-yellow-500 text-gray-50 inline-flex items-center justify-center text-3x w-10 h-10 rounded-full">
-                    W
-                  </span>
-                  <span className="mx-2">=</span>
-                  <span>Correct letter, wrong spot</span>
-                </div>
-                <span className="nm-inset-n-gray text-gray-50 inline-flex items-center justify-center text-3x w-10 h-10 rounded-full">
-                  W
-                </span>
-                <span className="mx-2">=</span>
-                <span>Wrong letter</span>
-                <label className="flex my-4 items-center">
-                  <input
-                    className="m-2"
-                    type="checkbox"
-                    checked={darkMode}
-                    onClick={toggleDarkMode}
-                  />
-                  <span>Dark Mode</span>
-                </label>
-              </div>
-              <div className="flex justify-center sm:text-base text-sm">
-                <span>This project is open source on</span>
-                <a
-                  className="ml-[6px] rounded-full h-5 w-5 sm:h-6 sm:w-6"
-                  href="https://github.com/octokatherine/word-master"
-                  target="_blank"
-                  rel="noreferrer"
-                >
-                  <Github />
-                </a>
-              </div>
-            </div>
-          </div>
-        </Modal>
       </div>
     </div>
   )
